refactoring - mail services
