+++ app/presenters/UzivatelActionsPresenter.php
@@ -4,6 +4,7 @@ namespace App\Presenters;
 
 use Nette,
     App\Model,
+    App\Services,
     Tracy\Debugger;
 
 /**
@@ -14,11 +15,13 @@ class UzivatelActionsPresenter extends UzivatelPresenter
     private $accountActivation;
     private $uzivatel;
     private $pdfGenerator;
+    private $mailService;
 
-    function __construct(Model\PdfGenerator $pdf, Model\AccountActivation $accActivation, Model\Uzivatel $uzivatel) {
+    function __construct(Services\MailService $mailsvc, Services\PdfGenerator $pdf, Model\AccountActivation $accActivation, Model\Uzivatel $uzivatel) {
         $this->pdfGenerator = $pdf;
         $this->accountActivation = $accActivation;
         $this->uzivatel = $uzivatel;
+        $this->mailService = $mailsvc;
     }
 
     public function actionMoneyActivate() {
@@ -69,6 +72,42 @@ class UzivatelActionsPresenter extends UzivatelPresenter
                 $pdftemplate = $this->createTemplate()->setFile(__DIR__."/../templates/Uzivatel/pdf-form.latte");
                 $pdf = $this->pdfGenerator->generatePdf($uzivatel, $pdftemplate);
                 $this->sendResponse($pdf);
+            }
+        }
+    }
+    public function actionSendRegActivation() {
+        if($this->getParam('id'))
+        {
+            if($uzivatel = $this->uzivatel->getUzivatel($this->getParam('id')))
+    	    {
+                $hash = base64_encode($uzivatel->id.'-'.md5($this->context->parameters["salt"].$uzivatel->zalozen));
+                $link = "https://moje.hkfree.org/uzivatel/confirm/".$hash;
+                //\Tracy\Dumper::dump($link);exit();
+                $so = $this->uzivatel->getUzivatel($this->getUser()->getIdentity()->getId());
+
+                $this->mailService->sendConfirmationRequest($uzivatel, $so, $link);
+                $this->mailService->sendConfirmationRequestCopy($uzivatel, $so);
+                
+                $this->flashMessage('E-mail s žádostí o potvrzení registrace byl odeslán.');
+
+                $this->redirect('Uzivatel:show', array('id'=>$uzivatel->id));
+            }
+        }
+    }
+
+    public function actionExportAndSendRegForm() {
+        if($this->getParam('id'))
+        {
+            if($uzivatel = $this->uzivatel->getUzivatel($this->getParam('id')))
+    	    {
+                $pdftemplate = $this->createTemplate()->setFile(__DIR__."/../templates/Uzivatel/pdf-form.latte");
+                $pdf = $this->pdfGenerator->generatePdf($uzivatel, $pdftemplate);
+
+                $this->mailService->mailPdf($pdf, $uzivatel, $this->getHttpRequest(), $this->getHttpResponse(), $this->getUser()->getIdentity()->getId());
+
+                $this->flashMessage('E-mail byl odeslán.');
+
+                $this->redirect('Uzivatel:show', array('id'=>$uzivatel->id));
             }
         }
     }
